refactor: move axios_config.tsx to new config dir
