+++ react-frontend/src/components/Upload.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import api from "./axios_config";
+import api from "../config/axios_config";
 import { useNavigate } from "react-router-dom";
 
 interface ListItemProps {
